Reorganized layout + routes
